Outlining unit_01 and unit_02
Selected a dataset for the first unit. Opted for a larger dataset with more categorical variables over the iris dataset. Some code included is just a fun little side quest, however, I have outline the different parts of the unit that can be developed and documented further at a later date.

Added some context to the readme, as it's quite barren atm. Included first iteration of the lesson list for unit 2.