Update unit_01.ipynb
Filled out subsections for the loading up data, defining of data types, and began the first part of the analysis of one categorical variable using sns and pandas to generate a pictograph. Allowed for the graphing to be flexible within a given input range. Next will be bar charts and pie charts, which hopefully should be a quicker implementation, as I had to make specific augmentations to make a sns.catplot() into a makeshift pictogram.