Several Updates
Still haven't found a remedy for the animations yet, so that's still up next for me.
I don't think I made any big differences to unit_01.ipynb in stats khan.
Brought in a nasa dataset on the global temperature index over the last couple centuries for line graphs.
Updated the histogram section along with shapes of distributions.
Added the first parts of the line graphs section. To be improved later.